requirements.txt
aiosqlite==0.19.0
alembic==1.10.4
anyio==3.6.2
apprise==1.3.0
asgi-lifespan==2.1.0
asyncpg==0.27.0
attrs==23.1.0
azure-common==1.1.28
azure-core==1.26.4
azure-identity==1.13.0
azure-mgmt-containerinstance==10.1.0
azure-mgmt-core==1.4.0
azure-mgmt-resource==23.0.1
azure-storage-blob==12.16.0
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
coolname==2.2.0
croniter==1.3.14
cryptography==40.0.2
databricks-connect==11.3.11
dateparser==1.1.8
docker==6.1.2
fastapi==0.95.1
fsspec==2023.5.0
google-auth==2.18.0
greenlet==2.0.2
griffe==0.27.4
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.0
httpx==0.24.0
hyperframe==6.0.1
idna==3.4
isodate==0.6.1
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.17.3
kubernetes==26.1.0
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
msal==1.22.0
msal-extensions==1.0.0
oauthlib==3.2.2
orjson==3.8.12
packaging==23.1
pathspec==0.11.1
pendulum==2.1.2
portalocker==2.7.0
prefect==2.10.9
prefect-azure==0.2.7
prefect-databricks==0.1.4
py4j==0.10.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.7
Pygments==2.15.1
PyJWT==2.7.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
PyYAML==6.0
readchar==4.0.5
regex==2023.5.5
requests==2.30.0
requests-oauthlib==1.3.1
rich==13.3.5
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.48
starlette==0.26.1
text-unidecode==1.3
toml==0.10.2
typer==0.9.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
uvicorn==0.22.0
websocket-client==1.5.1
websockets==11.0.3
